Logic refactoring and file organization.
