+++ app/Services/Report/FetchAllOrdersByDate.php
@@ -2,25 +2,30 @@
 
 namespace App\Services\Report;
 
+use App\Models\Type;
 use App\Models\Order;
-use App\Services\CachedService;
 use PerfectOblivion\Services\Traits\SelfCallingService;
 
-class FetchAllOrdersByDate extends CachedService
+class FetchAllOrdersByDate
 {
     use SelfCallingService;
 
     /** @var \App\Models\Order */
     private $orders;
 
+    /** @var \App\Models\Type */
+    private $types;
+
     /**
      * Construct a new FetchAllOrdersByDate service.
      *
      * @param  \App\Models\Order  $orders
+     * @param  \App\Models\Type  $types
      */
-    public function __construct(Order $orders)
+    public function __construct(Order $orders, Type $types)
     {
         $this->orders = $orders;
+        $this->types = $types;
     }
 
     /**
@@ -32,8 +37,8 @@ class FetchAllOrdersByDate extends CachedService
      */
     public function run(string $date)
     {
-        return resolve('orders')->map(function ($type) use ($date) {
-            return $this->orders->with('types')->type($type)->forDate($date)->get();
+        return $this->types->all()->keyBy('type')->map(function ($model, $key) use ($date) {
+            return $this->orders->with('types')->type($key)->forDate($date)->get();
         });
     }
 }
