+++ config/app.php
@@ -71,7 +71,6 @@ return [
         App\Providers\UrlServiceProvider::class,
         App\Providers\DebugbarServiceProvider::class,
         App\Providers\IdeHelperServiceProvider::class,
-        App\Providers\OrdersServiceProvider::class,
     ],
 
     'aliases' => [
